requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Azure Speech SDK for pronunciation assessment via SDK
azure-cognitiveservices-speech
